+++ backend/jobimport/test_dxf_parser.py
@@ -12,7 +12,7 @@ dxf_string = fin.read()
 fin.close()
 
 dxf_string = str(dxf_string)
-dxfParser = DXFParser(0.8)
+dxfParser = DXFParser(0.8, [1220, 610])
 forced_unit = 0
 
 parse_results = dxfParser.parse(dxf_string, forced_unit)